remove mock dependency from bank tests
