+++ x/bank/app_test.go
@@ -3,10 +3,11 @@ package bank_test
 import (
 	"testing"
 
+	dbm "github.com/tendermint/tm-db"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/auth"
 	"github.com/cosmos/cosmos-sdk/x/bank/internal/types"
-	"github.com/cosmos/cosmos-sdk/x/mock"
 
 	"github.com/stretchr/testify/require"
 
@@ -97,17 +98,20 @@ var (
 )
 
 func TestSendNotEnoughBalance(t *testing.T) {
-	mapp := getMockApp(t)
+	db := dbm.NewMemDB()
+	app := simapp.NewSimApp(log.NewTMLogger(log.NewSyncWriter(os.Stdout)), db, nil, true, 0)
+
 	acc := &auth.BaseAccount{
 		Address: addr1,
 		Coins:   sdk.Coins{sdk.NewInt64Coin("foocoin", 67)},
 	}
 
-	mock.SetGenesis(mapp, []auth.Account{acc})
+	// TODO update
+	mock.SetGenesis(app, []auth.Account{acc})
 
-	ctxCheck := mapp.BaseApp.NewContext(true, abci.Header{})
+	ctxCheck := app.BaseApp.NewContext(true, abci.Header{})
 
-	res1 := mapp.AccountKeeper.GetAccount(ctxCheck, addr1)
+	res1 := app.AccountKeeper.GetAccount(ctxCheck, addr1)
 	require.NotNil(t, res1)
 	require.Equal(t, acc, res1.(*auth.BaseAccount))
 
@@ -115,12 +119,12 @@ func TestSendNotEnoughBalance(t *testing.T) {
 	origSeq := res1.GetSequence()
 
 	sendMsg := types.NewMsgSend(addr1, addr2, sdk.Coins{sdk.NewInt64Coin("foocoin", 100)})
-	header := abci.Header{Height: mapp.LastBlockHeight() + 1}
-	mock.SignCheckDeliver(t, mapp.Cdc, mapp.BaseApp, header, []sdk.Msg{sendMsg}, []uint64{origAccNum}, []uint64{origSeq}, false, false, priv1)
+	header := abci.Header{Height: app.LastBlockHeight() + 1}
+	simapp.SignCheckDeliver(t, app.Codec(), app.BaseApp, header, []sdk.Msg{sendMsg}, []uint64{origAccNum}, []uint64{origSeq}, false, false, priv1)
 
-	mock.CheckBalance(t, mapp, addr1, sdk.Coins{sdk.NewInt64Coin("foocoin", 67)})
+	simapp.CheckBalance(t, app, addr1, sdk.Coins{sdk.NewInt64Coin("foocoin", 67)})
 
-	res2 := mapp.AccountKeeper.GetAccount(mapp.NewContext(true, abci.Header{}), addr1)
+	res2 := app.AccountKeeper.GetAccount(app.NewContext(true, abci.Header{}), addr1)
 	require.NotNil(t, res2)
 
 	require.True(t, res2.GetAccountNumber() == origAccNum)
@@ -128,7 +132,9 @@ func TestSendNotEnoughBalance(t *testing.T) {
 }
 
 func TestSendToModuleAcc(t *testing.T) {
-	mapp := getMockApp(t)
+	db := dbm.NewMemDB()
+	app := simapp.NewSimApp(log.NewTMLogger(log.NewSyncWriter(os.Stdout)), db, nil, true, 0)
+
 	acc := &auth.BaseAccount{
 		Address: addr1,
 		Coins:   coins,
@@ -140,22 +146,22 @@ func TestSendToModuleAcc(t *testing.T) {
 
 	mock.SetGenesis(mapp, []auth.Account{acc, macc})
 
-	ctxCheck := mapp.BaseApp.NewContext(true, abci.Header{})
+	ctxCheck := app.BaseApp.NewContext(true, abci.Header{})
 
-	res1 := mapp.AccountKeeper.GetAccount(ctxCheck, addr1)
+	res1 := app.AccountKeeper.GetAccount(ctxCheck, addr1)
 	require.NotNil(t, res1)
 	require.Equal(t, acc, res1.(*auth.BaseAccount))
 
 	origAccNum := res1.GetAccountNumber()
 	origSeq := res1.GetSequence()
 
-	header := abci.Header{Height: mapp.LastBlockHeight() + 1}
-	mock.SignCheckDeliver(t, mapp.Cdc, mapp.BaseApp, header, []sdk.Msg{sendMsg2}, []uint64{origAccNum}, []uint64{origSeq}, false, false, priv1)
+	header := abci.Header{Height: app.LastBlockHeight() + 1}
+	simapp.SignCheckDeliver(t, app.Codec(), app.BaseApp, header, []sdk.Msg{sendMsg2}, []uint64{origAccNum}, []uint64{origSeq}, false, false, priv1)
 
-	mock.CheckBalance(t, mapp, addr1, coins)
-	mock.CheckBalance(t, mapp, moduleAccAddr, sdk.Coins(nil))
+	simapp.CheckBalance(t, app, addr1, coins)
+	simapp.CheckBalance(t, app, moduleAccAddr, sdk.Coins(nil))
 
-	res2 := mapp.AccountKeeper.GetAccount(mapp.NewContext(true, abci.Header{}), addr1)
+	res2 := app.AccountKeeper.GetAccount(app.NewContext(true, abci.Header{}), addr1)
 	require.NotNil(t, res2)
 
 	require.True(t, res2.GetAccountNumber() == origAccNum)
@@ -163,7 +169,9 @@ func TestSendToModuleAcc(t *testing.T) {
 }
 
 func TestMsgMultiSendWithAccounts(t *testing.T) {
-	mapp := getMockApp(t)
+	db := dbm.NewMemDB()
+	app := simapp.NewSimApp(log.NewTMLogger(log.NewSyncWriter(os.Stdout)), db, nil, true, 0)
+
 	acc := &auth.BaseAccount{
 		Address: addr1,
 		Coins:   sdk.Coins{sdk.NewInt64Coin("foocoin", 67)},
@@ -175,9 +183,9 @@ func TestMsgMultiSendWithAccounts(t *testing.T) {
 
 	mock.SetGenesis(mapp, []auth.Account{acc, macc})
 
-	ctxCheck := mapp.BaseApp.NewContext(true, abci.Header{})
+	ctxCheck := app.BaseApp.NewContext(true, abci.Header{})
 
-	res1 := mapp.AccountKeeper.GetAccount(ctxCheck, addr1)
+	res1 := app.AccountKeeper.GetAccount(ctxCheck, addr1)
 	require.NotNil(t, res1)
 	require.Equal(t, acc, res1.(*auth.BaseAccount))
 
@@ -213,17 +221,18 @@ func TestMsgMultiSendWithAccounts(t *testing.T) {
 	}
 
 	for _, tc := range testCases {
-		header := abci.Header{Height: mapp.LastBlockHeight() + 1}
-		mock.SignCheckDeliver(t, mapp.Cdc, mapp.BaseApp, header, tc.msgs, tc.accNums, tc.accSeqs, tc.expSimPass, tc.expPass, tc.privKeys...)
+		header := abci.Header{Height: app.LastBlockHeight() + 1}
+		simapp.SignCheckDeliver(t, app.Codec(), app.BaseApp, header, tc.msgs, tc.accNums, tc.accSeqs, tc.expSimPass, tc.expPass, tc.privKeys...)
 
 		for _, eb := range tc.expectedBalances {
-			mock.CheckBalance(t, mapp, eb.addr, eb.coins)
+			simapp.CheckBalance(t, app, eb.addr, eb.coins)
 		}
 	}
 }
 
 func TestMsgMultiSendMultipleOut(t *testing.T) {
-	mapp := getMockApp(t)
+	db := dbm.NewMemDB()
+	app := simapp.NewSimApp(log.NewTMLogger(log.NewSyncWriter(os.Stdout)), db, nil, true, 0)
 
 	acc1 := &auth.BaseAccount{
 		Address: addr1,
@@ -253,17 +262,18 @@ func TestMsgMultiSendMultipleOut(t *testing.T) {
 	}
 
 	for _, tc := range testCases {
-		header := abci.Header{Height: mapp.LastBlockHeight() + 1}
-		mock.SignCheckDeliver(t, mapp.Cdc, mapp.BaseApp, header, tc.msgs, tc.accNums, tc.accSeqs, tc.expSimPass, tc.expPass, tc.privKeys...)
+		header := abci.Header{Height: app.LastBlockHeight() + 1}
+		simapp.SignCheckDeliver(t, app.Codec(), app.BaseApp, header, tc.msgs, tc.accNums, tc.accSeqs, tc.expSimPass, tc.expPass, tc.privKeys...)
 
 		for _, eb := range tc.expectedBalances {
-			mock.CheckBalance(t, mapp, eb.addr, eb.coins)
+			simapp.CheckBalance(t, app, eb.addr, eb.coins)
 		}
 	}
 }
 
 func TestMsgMultiSendMultipleInOut(t *testing.T) {
-	mapp := getMockApp(t)
+	db := dbm.NewMemDB()
+	app := simapp.NewSimApp(log.NewTMLogger(log.NewSyncWriter(os.Stdout)), db, nil, true, 0)
 
 	acc1 := &auth.BaseAccount{
 		Address: addr1,
@@ -298,17 +308,18 @@ func TestMsgMultiSendMultipleInOut(t *testing.T) {
 	}
 
 	for _, tc := range testCases {
-		header := abci.Header{Height: mapp.LastBlockHeight() + 1}
-		mock.SignCheckDeliver(t, mapp.Cdc, mapp.BaseApp, header, tc.msgs, tc.accNums, tc.accSeqs, tc.expSimPass, tc.expPass, tc.privKeys...)
+		header := abci.Header{Height: app.LastBlockHeight() + 1}
+		simapp.SignCheckDeliver(t, app.Codec(), app.BaseApp, header, tc.msgs, tc.accNums, tc.accSeqs, tc.expSimPass, tc.expPass, tc.privKeys...)
 
 		for _, eb := range tc.expectedBalances {
-			mock.CheckBalance(t, mapp, eb.addr, eb.coins)
+			simapp.CheckBalance(t, app, eb.addr, eb.coins)
 		}
 	}
 }
 
 func TestMsgMultiSendDependent(t *testing.T) {
-	mapp := getMockApp(t)
+	db := dbm.NewMemDB()
+	app := simapp.NewSimApp(log.NewTMLogger(log.NewSyncWriter(os.Stdout)), db, nil, true, 0)
 
 	acc1 := auth.NewBaseAccountWithAddress(addr1)
 	acc2 := auth.NewBaseAccountWithAddress(addr2)
@@ -347,10 +358,10 @@ func TestMsgMultiSendDependent(t *testing.T) {
 
 	for _, tc := range testCases {
 		header := abci.Header{Height: mapp.LastBlockHeight() + 1}
-		mock.SignCheckDeliver(t, mapp.Cdc, mapp.BaseApp, header, tc.msgs, tc.accNums, tc.accSeqs, tc.expSimPass, tc.expPass, tc.privKeys...)
+		simapp.SignCheckDeliver(t, app.Codec(), app.BaseApp, header, tc.msgs, tc.accNums, tc.accSeqs, tc.expSimPass, tc.expPass, tc.privKeys...)
 
 		for _, eb := range tc.expectedBalances {
-			mock.CheckBalance(t, mapp, eb.addr, eb.coins)
+			simapp.CheckBalance(t, app, eb.addr, eb.coins)
 		}
 	}
 }
