+++ x/mock/test_utils.go
@@ -67,42 +67,3 @@ func CheckGenTx(
 
 	return res
 }
-
-// SignCheckDeliver checks a generated signed transaction and simulates a
-// block commitment with the given transaction. A test assertion is made using
-// the parameter 'expPass' against the result. A corresponding result is
-// returned.
-func SignCheckDeliver(
-	t *testing.T, cdc *codec.Codec, app *baseapp.BaseApp, header abci.Header, msgs []sdk.Msg,
-	accNums, seq []uint64, expSimPass, expPass bool, priv ...crypto.PrivKey,
-) sdk.Result {
-
-	tx := GenTx(msgs, accNums, seq, priv...)
-
-	txBytes, err := cdc.MarshalBinaryLengthPrefixed(tx)
-	require.Nil(t, err)
-
-	// Must simulate now as CheckTx doesn't run Msgs anymore
-	res := app.Simulate(txBytes, tx)
-
-	if expSimPass {
-		require.Equal(t, sdk.CodeOK, res.Code, res.Log)
-	} else {
-		require.NotEqual(t, sdk.CodeOK, res.Code, res.Log)
-	}
-
-	// Simulate a sending a transaction and committing a block
-	app.BeginBlock(abci.RequestBeginBlock{Header: header})
-	res = app.Deliver(tx)
-
-	if expPass {
-		require.Equal(t, sdk.CodeOK, res.Code, res.Log)
-	} else {
-		require.NotEqual(t, sdk.CodeOK, res.Code, res.Log)
-	}
-
-	app.EndBlock(abci.RequestEndBlock{})
-	app.Commit()
-
-	return res
-}
